+++ src/main/java/br/com/cvc/scheduling/transfer/SchedulingTransferServiceApplication.java
@@ -1,0 +1,13 @@
+package br.com.cvc.scheduling.transfer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SchedulingTransferServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SchedulingTransferServiceApplication.class, args);
+	}
+
+}
