--- src/main/java/br/com/cvc/scheduling/transfer/service/SchedulingTransferServiceApplication.java
@@ -1,13 +1,0 @@
-package br.com.cvc.scheduling.transfer.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SchedulingTransferServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SchedulingTransferServiceApplication.class, args);
-	}
-
-}
