+++ src/test/java/br/com/cvc/scheduling/transfer/SchedulingTransferServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.cvc.scheduling.transfer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SchedulingTransferServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
