--- src/test/java/br/com/cvc/scheduling/transfer/service/SchedulingTransferServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.cvc.scheduling.transfer.service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SchedulingTransferServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
