--- src/test/java/br/com/cvc/scheduling/transfer/service/util/FailingRunnable.java
@@ -1,3 +1,0 @@
-package br.com.cvc.scheduling.transfer.service.util;
-
-public interface FailingRunnable { void run() throws Exception; }
