+++ src/test/java/br/com/cvc/scheduling/transfer/util/ControllerTest.java
@@ -1,0 +1,25 @@
+package br.com.cvc.scheduling.transfer.util;
+
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target({TYPE, ANNOTATION_TYPE})
+@Retention(RUNTIME)
+@Tag("controller")
+@ExtendWith(SpringExtension.class)
+@WebMvcTest
+public @interface ControllerTest {
+
+    @AliasFor(annotation = WebMvcTest.class, attribute = "controllers")
+    Class<?>[] value() default {};
+}
