+++ src/test/java/br/com/cvc/scheduling/transfer/util/FailingRunnable.java
@@ -1,0 +1,3 @@
+package br.com.cvc.scheduling.transfer.util;
+
+public interface FailingRunnable { void run() throws Exception; }
